requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
altair=4.2.0=pypi_0
astunparse=1.6.3=pypi_0
attrs=22.2.0=pypi_0
blinker=1.5=pypi_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py39h06a4308_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
commonmark=0.9.1=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
flatbuffers=22.12.6=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=5.2.0=pypi_0
install=1.3.5=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=14.0.6=pypi_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.2=pypi_0
ncurses=6.3=h5eee18b_3
numpy=1.24.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=22.0=pypi_0
pandas=1.5.2=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=py39h06a4308_0
protobuf=3.19.6=pypi_0
pyarrow=10.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.13.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.2=pypi_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytz=2022.7=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=12.6.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.0=pypi_0
scipy=1.9.3=pypi_0
semver=2.13.0=pypi_0
setuptools=65.5.0=py39h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.40.0=h5082296_0
streamlit=1.16.0=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.29.0=pypi_0
termcolor=2.1.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.13=pypi_0
validators=0.20.0=pypi_0
watchdog=2.2.0=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.8=h5eee18b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
